chore: refactor NestJS file structure
- apply modular architecture to file structure
