+++ backend/src/common/enums/address-entity.enum.ts
@@ -1,0 +1,5 @@
+export enum AddressEntityType {
+  PATIENT = 'Patient',
+  USER = 'User',
+  EMPLOYEE = 'Employee',
+}
