+++ backend/src/modules/domains/extraction/extraction.constants.ts
