+++ backend/src/modules/domains/extraction/utils/extraction.utils.ts
