+++ backend/src/modules/domains/patients/entities/consultation.entity.ts
@@ -8,7 +8,7 @@ import {
   ManyToOne,
   JoinColumn,
 } from 'typeorm';
-import { Patient } from '../../patients/entities/patient.entity';
+import { Patient } from './patient.entity';
 
 /**
  * Represents the vitals taken during a consultation.
