+++ backend/src/modules/domains/patients/utils/patient-formatting.utils.ts
