+++ backend/src/modules/domains/users/entities/user.entity.ts
